Added

+++ cbio.fix.data_expression.sh
@@ -1,13 +1,19 @@
-for i in */tcga/data_expression.txt;
- do A=`dirname $i`;
-    cd $A;
-    echo '[WORKING ON] ' $i;
-    pwd;
-    cp data_expression.txt data_expression.BAK;
-    more data_expression.txt | grep $'Composite Element REF\t\tSignal' -v > a;
-    mv a data_expression.txt
+for i in */tcga/;
+ do echo '[DATA_EXPRESSION.txt] : ' $i;
+    cd $i;
+    if [ -e data_expression.txt ]
+    then       
+	echo '> [RUNNING]';
+	cp data_expression.txt data_expression.BAK;
+	more data_expression.txt | grep $'Composite Element REF\t\tSignal' -v > a;
+	mv a data_expression.txt
+    else
+	echo '> [SKIP] : does not exists';
+	
+    fi
     cd ../../;
- done
+    echo ''
+done
